--- SoftLayer/CLI/rwhois/__init__.py
@@ -1,1 +1,0 @@
-"""Referral Whois."""
