--- docs/cli/rwhois.rst
@@ -1,12 +1,0 @@
-.. _cli_rwhois:
-
-Reverse Whois Commands
-======================
-
-.. click:: SoftLayer.CLI.rwhois.edit:cli
-    :prog: rwhois edit
-    :show-nested:
-
-.. click:: SoftLayer.CLI.rwhois.show:cli
-    :prog: rwhois show
-    :show-nested:
